separation between skipNextLine and skipNextNullLine
to skip null termination after a string line (strncpy) the "last
line that was parsed" is stored in lastNonSkippedLine. then
if skipNextNullLine is set and if the current line has ' = 0;'
the current line is skipped.

still not optimal to catch all null termination after string lines,
yet all string lines seem to have null termination after them.

these recent changes in canSkip() makes it so that the parser now skips
both null termination and strupr lines after the actual string line
(with strncpy).