add support for defining template sizes
adding the following on top of the input file:
#define_sz <decimal>

will set a template size. the value for <decimal> will be checked against
the resulted template size. if the resulted value is smaller, auto-padding
will be added at the end.

a cosmetic change is that the hex value of <decimal> will
be placed next to the class name if #define_sz is used.