add the stringNullFixup1 (define_no_string_null_fixup1) toggle
most strings seem to be null terminated at SZ, and the next
property starts at the correct offset.

but there are cases where one byte of padding is left after
the null terminating byte, which is a bit difficult to handle.
we can either adjust the size of those strings by 1 manually
or add one byte of padding.

show a warning in those cases and add a way to toggle the
behavior with 'define_no_string_null_fixup1'.

note that if 'define_no_string_null_fixup1' is not set
the total template size might end up being wrong.